tests/requirements.txt
gitpython==3.1.7
html2markdown==0.1.7
humanfriendly==9.1
humanize==2.5.0
mock==4.0.2
pytest==6.0.0
pytest-xdist==1.34.0
junitparser==1.4.1
gitpython==3.1.7
